requirements.txt
android-backup==0.2.0
appdirs==1.4.4
async-timeout==4.0.2
attrs==22.1.0
backports.zoneinfo==0.2.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
construct==2.10.68
croniter==1.3.8
cryptography==38.0.4
defusedxml==0.7.1
idna==3.4
ifaddr==0.2.0
micloud==0.6
netifaces==0.11.0
pycparser==2.21
pycryptodome==3.16.0
pygame==2.1.2
python-dateutil==2.8.2
python-dotenv==0.21.0
python-miio==0.5.12
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests==2.28.1
six==1.16.0
tqdm==4.64.1
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.13
zeroconf==0.39.4
